Added Schrodinger Eq tutorial
Added a tutorial for the Schrodinger Equation.  I moved all the main folders into a qudipy folder in the repo.  This follows standard library conventions (see numpy/scipy repos). There is a .ipyn folder that was added, which I'm not sure if I can remove or not.. So i'm keeping it for now.  But perhaps we should add that to the .gitignore if it doesn't prevent us from viewing the tutorials in the repo.